bank-pipeline/requirements.txt
pyspark==3.5.1
pyyaml>=6
python-dateutil>=2.9
pandas>=2
pytest>=8
